implemented issue CS-2135 : Improve the performance of list articles display
Created a new stage in the article workflow: "publish with issue". Articles
from unpublished issues are set to this stage for publishing. When the issue
was published the article stage is set to published. When the issue was
unpublished the published article stage is set to "publish with issue". This
way there is no need to create a join with the Issues table when listing
published articles (articles that are in the publish stage and the issue was
also published).