Implemented issue CS-2143 : Optimize and improve the template error reporting
When encountering a critical error the template engine does not display any
message so the user does not know there is an error in the template syntax.
E.g.: $campsite->article->type or $campsite->issue generates an error in PHP
as the return is an object that can't be displayed. Campsite should report
this as invalid use of the template language.

Modified CampRequest to use a global variable because depending on the context
it was included from sometimes the static member has to be reinitialized.

Use $GLOBALS['var'] wherever we access global variables instead of declaring
the global var in the function and using it as a local var.

Fixed language setting for the error reporting in the preview window.