CS-3090 : Implement the themes list screen
Added json encoding and quoting for jquery utility helper register var
Added a pass to duplicate key exception in the theme management service with a TODO to remove if the logic is not correct
i.e. Row added for theme - publication assignation, files missing -> should keep the row and just copy the files there.